fix(tests): resolve meshcore_feeder fixture double-registration
Move meshcore_feeder to nodes.tests.conftest (root pytest_plugins) so
meshcore_packets/tests/conftest is not loaded twice. Keeps MC constants
in meshcore_packets for URL helpers.
